Split off SidebarButton
