+++ osu.Game/osu.Game.csproj
@@ -234,6 +234,7 @@
     <Compile Include="Overlays\Options\Online\PrivacyOptions.cs" />
     <Compile Include="Overlays\Options\Online\NotificationsOptions.cs" />
     <Compile Include="Overlays\Options\CheckBoxOption.cs" />
+    <Compile Include="Overlays\Options\SidebarButton.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="$(SolutionDir)\osu-framework\osu.Framework\osu.Framework.csproj">
